[M12i] Address review feedback on force-push confirmation
- Propagate get_current_branch error instead of unwrap_or_default
- Detect non-interactive stdin and bail with clear message
- Extract is_affirmative to shared prompt_utils module
- Add requires = "push" to --yes flag in clap metadata
